+++ CHANGELOG.md
@@ -9,6 +9,12 @@ for further information about branching and tagging conventions.
 ## [Unreleased]
 #### Fixed
 - A bug with localization
+- A bug with incorrect display of the amount of the maximum issued asset in
+  assetRequest.Show
+#### Changed
+- helper localize
+#### Added
+- localize to issuedAmount in assetRequest.Show 
 
 ## [1.12.0-rc.3] - 2020-02-18
 #### Fixed